# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/service-request": {
      "post": {
        "operationId": "ServiceRequestController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "service-request"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ServiceRequestController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "service-request"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/service-request/{ids}": {
      "get": {
        "operationId": "ServiceRequestController_findOne",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "service-request"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/service-request/srUpdate/{id}": {
      "patch": {
        "operationId": "ServiceRequestController_updateServiceRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSRLifterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "service-request"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/roster": {
      "post": {
        "operationId": "RosterController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRosterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "roster"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "RosterController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roster"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/roster/{id}": {
      "get": {
        "operationId": "RosterController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roster"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "RosterController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRosterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roster"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "RosterController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roster"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "BENTECH",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateServiceRequestDto": {
        "type": "object",
        "properties": {
          "requestedTime": {
            "format": "date-time",
            "type": "string",
            "example": "2024-07-10T17:43:51.189Z"
          },
          "requestedQty": {
            "type": "number",
            "example": "5000"
          },
          "Type": {
            "type": "string",
            "enum": [
              "fluid",
              "solid",
              "semi-solid"
            ]
          }
        },
        "required": [
          "requestedTime",
          "requestedQty",
          "Type"
        ]
      },
      "UpdateSRLifterDto": {
        "type": "object",
        "properties": {
          "lifterId": {
            "type": "number",
            "example": 1
          },
          "rosterId": {
            "type": "number",
            "example": 1
          },
          "pickupTime": {
            "format": "date-time",
            "type": "string",
            "example": "2024-07-10T17:43:51.189Z"
          },
          "pickupQty": {
            "type": "number",
            "example": "5000"
          },
          "sendReport": {
            "type": "boolean",
            "example": true
          },
          "lifterFeedback": {
            "type": "string",
            "example": "Type your Feedback"
          },
          "Status": {
            "type": "string",
            "example": "completed",
            "enum": [
              "pending",
              "completed",
              "inprogress",
              "assign",
              "cancelled"
            ]
          }
        },
        "required": [
          "lifterId",
          "rosterId",
          "pickupTime",
          "pickupQty",
          "sendReport",
          "lifterFeedback",
          "Status"
        ]
      },
      "CreateRosterDto": {
        "type": "object",
        "properties": {
          "serviceRequestId": {
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assignedTime": {
            "format": "date-time",
            "type": "string",
            "default": "2024-07-10T17:43:51.191Z"
          },
          "location": {
            "type": "string",
            "example": "Mohali"
          },
          "vehicleNo": {
            "type": "string",
            "example": "PB01-33-3333"
          },
          "lifterId": {
            "type": "number",
            "example": 1
          },
          "Status": {
            "type": "string",
            "example": "completed",
            "enum": [
              "pending",
              "completed",
              "inprogress",
              "assign",
              "cancelled"
            ]
          }
        },
        "required": [
          "serviceRequestId",
          "assignedTime",
          "location",
          "vehicleNo",
          "lifterId",
          "Status"
        ]
      },
      "UpdateRosterDto": {
        "type": "object",
        "properties": {
          "assignedTime": {
            "format": "date-time",
            "type": "string",
            "default": "2024-07-10T17:43:51.191Z"
          },
          "vehicleNo": {
            "type": "string",
            "example": "PB01-33-3333"
          },
          "Status": {
            "type": "string",
            "example": "completed",
            "enum": [
              "pending",
              "completed",
              "inprogress",
              "assign",
              "cancelled"
            ]
          }
        },
        "required": [
          "assignedTime",
          "vehicleNo",
          "Status"
        ]
      }
    }
  }
}